Move code out of cli-args
